requirements.txt
fitz
whoosh
torch
faiss-cpu
transformers
streamlit
